Moved email settings out of the settings quarantine
no ref

The email settings components and the shared email-design pieces now live
in settings/email and settings/email-design, beside the email acceptance
tests. Moving them enables the type-aware lint rules the rest of admin
uses, so the same debt is cleared for these files: async handlers passed
to void callbacks are wrapped, fire-and-forget promises are marked, the
search keywords move into their own modules, and the email design context
is split the way the other admin contexts are — the context and hook in a
.ts module and the provider component in its own file — so each file only
exports one kind of thing.
