+++ apps/admin/src/settings/app/components/sidebar.tsx
@@ -10,8 +10,8 @@ import {useFocusContext} from '@tryghost/shade/app';
 import {checkStripeEnabled, getSettingValues} from '@tryghost/admin-x-framework/api/settings';
 
 import {searchKeywords as advancedSearchKeywords} from './settings/advanced/advanced-settings';
-import {searchKeywords as emailSearchKeywords} from './settings/email/email-settings';
-import {searchKeywords as emailsSearchKeywords} from './settings/email/emails';
+import {searchKeywords as emailSearchKeywords} from '@/settings/email/search-keywords';
+import {searchKeywords as emailsSearchKeywords} from '@/settings/email/emails-search-keywords';
 import {searchKeywords as generalSearchKeywords} from '@/settings/general/search-keywords';
 import {searchKeywords as growthSearchKeywords} from '@/settings/growth/search-keywords';
 import {searchKeywords as membershipSearchKeywords} from '@/settings/membership/search-keywords';
