+++ apps/admin/src/settings/email-design/design-fields/body-font-field.tsx
@@ -1,6 +1,6 @@
 import {FONT_OPTIONS} from './font-constants';
 import {Select, SelectContent, SelectItem, SelectTrigger, SelectValue} from '@tryghost/shade/components';
-import {useEmailDesign} from '@/settings/app/components/settings/email-design/email-design-context';
+import {useEmailDesign} from '@/settings/email-design/email-design-context';
 
 export const BodyFontField = () => {
     const {settings, onSettingsChange} = useEmailDesign();
