+++ apps/admin/src/settings/email-design/design-fields/button-color-field.tsx
@@ -1,8 +1,8 @@
 import ColorPickerField from '@/settings/app/components/color-picker-field';
 import {getAutoSwatchHex} from './color-swatch-helpers';
-import {useEmailDesign} from '@/settings/app/components/settings/email-design/email-design-context';
+import {useEmailDesign} from '@/settings/email-design/email-design-context';
 
-export const LinkColorField = () => {
+export const ButtonColorField = () => {
     const {settings, onSettingsChange, accentColor} = useEmailDesign();
     const autoSwatchHex = getAutoSwatchHex(settings.background_color);
 
@@ -21,9 +21,9 @@ export const LinkColorField = () => {
                     hex: autoSwatchHex
                 }
             ]}
-            title="Link color"
-            value={settings.link_color}
-            onChange={color => onSettingsChange({link_color: color})}
+            title="Button color"
+            value={settings.button_color}
+            onChange={color => onSettingsChange({button_color: color})}
         />
     );
 };
