+++ apps/admin/src/settings/email-design/design-fields/heading-weight-field.tsx
@@ -1,6 +1,6 @@
 import {FONT_WEIGHT_OPTIONS, getValidWeight} from './font-constants';
 import {Select, SelectContent, SelectItem, SelectTrigger, SelectValue} from '@tryghost/shade/components';
-import {useEmailDesign} from '@/settings/app/components/settings/email-design/email-design-context';
+import {useEmailDesign} from '@/settings/email-design/email-design-context';
 
 export const HeadingWeightField = () => {
     const {settings, onSettingsChange} = useEmailDesign();
