+++ apps/admin/src/settings/email-design/email-design-provider.tsx
@@ -1,0 +1,12 @@
+import {type ReactNode} from 'react';
+import {EmailDesignContext, type EmailDesignContextValue} from './email-design-context';
+
+interface EmailDesignProviderProps extends EmailDesignContextValue {
+    children: ReactNode;
+}
+
+export const EmailDesignProvider = ({settings, onSettingsChange, accentColor, children}: EmailDesignProviderProps) => (
+    <EmailDesignContext.Provider value={{settings, onSettingsChange, accentColor}}>
+        {children}
+    </EmailDesignContext.Provider>
+);
