+++ apps/admin/src/settings/email-design/types.ts
