+++ apps/admin/src/settings/email-design/welcome-email-preview-content.tsx
