+++ apps/admin/src/settings/email/search-keywords.ts
@@ -1,0 +1,6 @@
+export const searchKeywords = {
+    enableNewsletters: ['emails', 'newsletters', 'newsletter sending', 'enable', 'disable', 'turn on', 'turn off'],
+    newsletters: ['newsletters', 'emails', 'design', 'customization'],
+    defaultRecipients: ['newsletters', 'default recipients', 'emails'],
+    mailgun: ['mailgun', 'emails', 'newsletters']
+};
