+++ apps/admin/src/settings/routes.tsx
@@ -27,8 +27,8 @@ export const settingsRouteChildren: RouteObject[] = [
     {path: 'tiers/add', lazy: lazyComponent(() => import('./membership/tiers/tier-detail-modal'))},
     {path: 'tiers/:tierId', lazy: lazyComponent(() => import('./membership/tiers/tier-detail-modal'))},
     {path: 'stripe-connect', lazy: lazyComponent(() => import('./membership/stripe/stripe-connect-modal'))},
-    {path: 'newsletters/new', lazy: lazyComponent(() => import('./app/components/settings/email/newsletters/add-newsletter-modal'))},
-    {path: 'newsletters/:newsletterId', lazy: lazyComponent(() => import('./app/components/settings/email/newsletters/newsletter-detail-modal'))},
+    {path: 'newsletters/new', lazy: lazyComponent(() => import('./email/newsletters/add-newsletter-modal'))},
+    {path: 'newsletters/:newsletterId', lazy: lazyComponent(() => import('./email/newsletters/newsletter-detail-modal'))},
     {path: 'history/view/:userId?', lazy: lazyComponent(() => import('./app/components/settings/advanced/history-modal'))},
     {path: 'integrations/new', lazy: lazyComponent(() => import('./app/components/settings/advanced/integrations/add-integration-modal'))},
     {path: 'integrations/contentapi', lazy: lazyComponent(() => import('./app/components/settings/advanced/integrations/content-api-modal'))},
